etcdctl: modify etcdctl v2 and v3 for code coverage
